Refactoring into walkable utils done
