+++ src/systems/vox-system.js
@@ -13,7 +13,6 @@ import { addMedia, isLockedMedia, addMediaInFrontOfPlayerIfPermitted } from "../
 import { ensureOwnership } from "../utils/ownership-utils";
 import { Voxels, rgbtForVoxColor, REMOVE_VOXEL_COLOR, ModelWriter, Color } from "smoothvoxels";
 import { SvoxBufferGeometry } from "smoothvoxels/three";
-import { generateAllWalkableMeshesWithinXZ } from "../utils/walk-utils";
 import VoxMesherWorker from "../workers/vox-mesher.worker.js";
 import {
   CompressionStream as CompressionStreamImpl,
@@ -782,7 +781,6 @@ export class VoxSystem extends EventTarget {
       pendingVoxels,
       pendingVoxelsOffset,
       showVoxGeometry,
-      pendingVoxMeshWorkerJobIds,
       hasWalkableSources
     } = entry;
 
@@ -1732,81 +1730,6 @@ export class VoxSystem extends EventTarget {
     return !entry.shapeIsEnvironmental;
   }
 
-  // Efficiently raycast to the closest walkable source, skipping non-walkable
-  // sources and avoiding extra raycasts. Only can cast up or down, so we can use
-  // bounding box to quickly cull as well.
-  raycastVerticallyToClosestWalkableSource = (function() {
-    const tmpMesh = new Mesh();
-    const instanceLocalMatrix = new Matrix4();
-    const instanceWorldMatrix = new Matrix4();
-    const instanceIntersects = [];
-    const raycaster = new THREE.Raycaster();
-    raycaster.firstHitOnly = true; // flag specific to three-mesh-bvh
-    raycaster.near = 0.01;
-    raycaster.far = 40;
-    raycaster.ray.direction.set(0, 1, 0);
-
-    return function(origin, up = true, backSide = false) {
-      const { voxIdToEntry } = this;
-
-      const { side } = voxMaterial;
-      let intersection = null;
-
-      if (backSide) {
-        voxMaterial.side = THREE.BackSide;
-      }
-
-      raycaster.ray.origin.copy(origin);
-      raycaster.ray.direction.y = up ? 1 : -1;
-
-      for (const entry of voxIdToEntry.values()) {
-        if (!entry.hasWalkableSources) continue;
-        const voxMesh = entry.meshes[0];
-        if (voxMesh === null) continue;
-
-        const { sources, walkableSources, walkGeometry } = entry;
-        if (walkGeometry === null) continue;
-
-        for (let instanceId = 0, l = sources.length; instanceId < l; instanceId++) {
-          const source = sources[instanceId];
-          if (source === null) continue;
-          if (!walkableSources[instanceId]) continue;
-
-          // Bounding box check for origin X,Z since we are casting up/down
-          const bbox = this.getBoundingBoxForSource(source, true);
-
-          if (origin.x < bbox.min.x || origin.x > bbox.max.x || origin.z < bbox.min.z || origin.z > bbox.max.z)
-            continue;
-
-          // Raycast once for each walkable source.
-          tmpMesh.geometry = walkGeometry;
-          tmpMesh.material = voxMaterial;
-          voxMesh.updateMatrices();
-          voxMesh.getMatrixAt(instanceId, instanceLocalMatrix);
-          instanceWorldMatrix.multiplyMatrices(voxMesh.matrixWorld, instanceLocalMatrix);
-          tmpMesh.matrixWorld = instanceWorldMatrix;
-          tmpMesh.raycast(raycaster, instanceIntersects);
-
-          if (instanceIntersects.length === 0) continue;
-
-          const newIntersection = instanceIntersects[0];
-
-          if (intersection === null || intersection.distance > newIntersection.distance) {
-            intersection = newIntersection;
-            intersection.instanceId = instanceId;
-            intersection.object = voxMesh;
-          }
-
-          instanceIntersects.length = 0;
-        }
-      }
-
-      voxMaterial.side = side;
-
-      return intersection;
-    };
-  })();
-
   generateWalkableMeshesWithinXZ = (function() {
     const tmpMesh = new Mesh();
     const instanceLocalMatrix = new Matrix4();
@@ -1849,66 +1772,6 @@ export class VoxSystem extends EventTarget {
 
           yield tmpMesh;
         }
-      }
-    };
-  })();
-
-  // Efficiently raycast to search for walls at knee-height of all walkable sources.
-  // sources and avoiding extra raycasts. Casts along a walk direction.
-  //
-  // We sample several ray directions to basically "sweep" the walk direction looking for any
-  // surfaces that could block the walk direction. This is necessary for things like stone walls
-  // that have crevices.
-  //
-  // For each surface we keep projecting a new walk direction onto that plane, carving down the
-  // possible walk direction that is compatible with all of them.
-  raycastForWallCheckToClosestWalkableSource = (function() {
-    const instanceIntersects = [];
-    const raycaster = new THREE.Raycaster();
-    const normalizedWalkDirection = new THREE.Vector3();
-    const finalWalkDirection = new THREE.Vector3();
-    const worldFaceNormal = new THREE.Vector3();
-    raycaster.firstHitOnly = true; // flag specific to three-mesh-bvh
-    raycaster.near = 0.01;
-    raycaster.far = 0.75; // Add a little buffer for avatar body + cel shading
-
-    return function(origin, walkDirection) {
-      let sawIntersection = false;
-      normalizedWalkDirection.copy(walkDirection).normalize();
-      finalWalkDirection.copy(walkDirection).normalize();
-      raycaster.ray.origin.copy(origin);
-
-      for (const mesh of generateAllWalkableMeshesWithinXZ(origin, 1, 1)) {
-        for (let dX = -0.5; dX <= 0.5; dX += 0.5) {
-          for (let dZ = -0.5; dZ <= 0.5; dZ += 0.5) {
-            raycaster.ray.direction.copy(normalizedWalkDirection);
-            raycaster.ray.direction.x += raycaster.ray.direction.x * dX;
-            raycaster.ray.direction.z += raycaster.ray.direction.z * dZ;
-
-            raycaster.ray.direction.normalize();
-
-            mesh.raycast(raycaster, instanceIntersects);
-
-            if (instanceIntersects.length === 0) continue;
-
-            sawIntersection = true;
-
-            for (const intersection of instanceIntersects) {
-              worldFaceNormal.copy(intersection.face.normal);
-              intersection.object.updateMatrices();
-              worldFaceNormal.transformDirection(intersection.object.matrixWorld);
-
-              if (worldFaceNormal.dot(normalizedWalkDirection) < 0) {
-                finalWalkDirection.projectOnPlane(worldFaceNormal);
-                finalWalkDirection.normalize();
-              }
-            }
-
-            instanceIntersects.length = 0;
-          }
-        }
-
-        return sawIntersection ? finalWalkDirection : null;
       }
     };
   })();
